refactor: move modals into a folder
